Added Foxy plugin into composer plugins list
Foxy is a plugin that acts like a bridge between npm/yarn. Once composer package is installed foxy checks for npm and yarn dependencies and if it founds something it tries to update all npm assets. 

I think it should be included in this awesome list about composer, so here is my pull request.